Fix various compiler warnings

+++ panda/src/display/graphicsStateGuardian.cxx
@@ -1954,7 +1954,7 @@ fetch_specified_texture(Shader::ShaderTexSpec &spec, SamplerState &sampler,
     {
       const TextureAttrib *texattrib;
       if (_target_rs->get_attrib(texattrib)) {
-        size_t si = 0;
+        int si = 0;
         for (int i = 0; i < texattrib->get_num_on_stages(); ++i) {
           TextureStage *stage = texattrib->get_on_stage(i);
           TextureStage::Mode mode = stage->get_mode();
@@ -1977,7 +1977,7 @@ fetch_specified_texture(Shader::ShaderTexSpec &spec, SamplerState &sampler,
     {
       const TextureAttrib *texattrib;
       if (_target_rs->get_attrib(texattrib)) {
-        size_t si = 0;
+        int si = 0;
         for (int i = 0; i < texattrib->get_num_on_stages(); ++i) {
           TextureStage *stage = texattrib->get_on_stage(i);
           TextureStage::Mode mode = stage->get_mode();
@@ -2005,7 +2005,7 @@ fetch_specified_texture(Shader::ShaderTexSpec &spec, SamplerState &sampler,
     {
       const TextureAttrib *texattrib;
       if (_target_rs->get_attrib(texattrib)) {
-        size_t si = 0;
+        int si = 0;
         for (int i = 0; i < texattrib->get_num_on_stages(); ++i) {
           TextureStage *stage = texattrib->get_on_stage(i);
           TextureStage::Mode mode = stage->get_mode();
@@ -2034,7 +2034,7 @@ fetch_specified_texture(Shader::ShaderTexSpec &spec, SamplerState &sampler,
     {
       const TextureAttrib *texattrib;
       if (_target_rs->get_attrib(texattrib)) {
-        size_t si = 0;
+        int si = 0;
         for (int i = 0; i < texattrib->get_num_on_stages(); ++i) {
           TextureStage *stage = texattrib->get_on_stage(i);
           TextureStage::Mode mode = stage->get_mode();
@@ -2057,7 +2057,7 @@ fetch_specified_texture(Shader::ShaderTexSpec &spec, SamplerState &sampler,
     {
       const TextureAttrib *texattrib;
       if (_target_rs->get_attrib(texattrib)) {
-        size_t si = 0;
+        int si = 0;
         for (int i = 0; i < texattrib->get_num_on_stages(); ++i) {
           TextureStage *stage = texattrib->get_on_stage(i);
           TextureStage::Mode mode = stage->get_mode();
@@ -2086,7 +2086,7 @@ fetch_specified_texture(Shader::ShaderTexSpec &spec, SamplerState &sampler,
     {
       const TextureAttrib *texattrib;
       if (_target_rs->get_attrib(texattrib)) {
-        size_t si = 0;
+        int si = 0;
         for (int i = 0; i < texattrib->get_num_on_stages(); ++i) {
           TextureStage *stage = texattrib->get_on_stage(i);
           TextureStage::Mode mode = stage->get_mode();
@@ -2107,7 +2107,7 @@ fetch_specified_texture(Shader::ShaderTexSpec &spec, SamplerState &sampler,
     {
       const TextureAttrib *texattrib;
       if (_target_rs->get_attrib(texattrib)) {
-        size_t si = 0;
+        int si = 0;
         for (int i = 0; i < texattrib->get_num_on_stages(); ++i) {
           TextureStage *stage = texattrib->get_on_stage(i);
           TextureStage::Mode mode = stage->get_mode();
